refactor: Update TaskStatusValidator namespace and adjust UserId requirement in CreateTaskEventDto; change StageType to required in CreateWorkflowStageDto
