+++ server/Server/Models/DTOs/TaskEvent/CreateTaskEventDto.cs
@@ -13,7 +13,6 @@ public record class CreateTaskEventDto
     [Required]
     public int TaskId { get; init; }
 
-    [Required]
     public string? UserId { get; init; }
 
     public int? FromWorkflowStageId { get; init; }
